requirements.txt
torch
transformers
discord.py
python-dotenv
SentencePiece
asyncio